Add config options (#124)
Not all config options were represented in the Android client. This PR adds them, along with sub-classes of `EventManagerConfig` to allow those modes have the necessary config.

Further work is still required on exposing EM mode callsbacks and functions, these will be in a separate PR.

This PR allows the Prompts API to be used, with support for JS callbacks within Java callbacks.

Co-authored-by: Steve Chaloner <steve@seats.io>